Initial chart label mixin/setup
