+++ addon/components/horizontal-bar-chart.js
@@ -6,8 +6,10 @@ import FloatingTooltipMixin from '../mixins/floating-tooltip';
 import SortableChartMixin from '../mixins/sortable-chart';
 
 import LabelTrimmer from '../utils/label-trimmer';
+import AxisTitles from '../mixins/axis-titles';
 
-export default ChartComponent.extend(FloatingTooltipMixin, FormattableMixin, SortableChartMixin, {
+export default ChartComponent.extend(FloatingTooltipMixin,
+  FormattableMixin, SortableChartMixin, AxisTitles, {
   classNames: ['chart-horizontal-bar'],
 
   // ----------------------------------------------------------------------------
@@ -258,6 +260,7 @@ export default ChartComponent.extend(FloatingTooltipMixin, FormattableMixin, Sor
   drawChart: function() {
     this.updateData();
     this.updateAxes();
+    this.updateAxisTitles();
     this.updateGraphic();
   },
