+++ tests/dummy/app/templates/horizontal-bar.hbs
@@ -10,6 +10,8 @@
         maxBarThickness=maxBarThickness
         selectedSeedColor=selectedSeedColor
         data=data
+        xValueDisplayName=xValueDisplayName
+        yValueDisplayName=yValueDisplayName
       }}
     </div>
   </div>
