+++ tests/dummy/app/templates/time-series.hbs
@@ -17,6 +17,8 @@
         yAxisFromZero=yAxisFromZero
         marginLeft=30
         dynamicXAxis=dynamicXAxis
+        xValueDisplayName=xValueDisplayName
+        yValueDisplayName=yValueDisplayName
       }}
     </div>
   </div>
