+++ tests/dummy/app/templates/vertical-bar.hbs
@@ -13,6 +13,8 @@
         withinGroupPadding=withinGroupPadding
         betweenGroupPadding=betweenGroupPadding
         stackBars=stackBars
+        xValueDisplayName=xValueDisplayName
+        yValueDisplayName=yValueDisplayName
       }}
     </div>
   </div>
