+++ tests/unit/mixins/axis-titles-test.js
@@ -1,0 +1,12 @@
+import Ember from 'ember';
+import AxisTitlesMixin from '../../../mixins/axis-titles';
+import { module, test } from 'qunit';
+
+module('Unit | Mixin | axis titles');
+
+// Replace this with your real tests.
+test('it works', function(assert) {
+  var AxisTitlesObject = Ember.Object.extend(AxisTitlesMixin);
+  var subject = AxisTitlesObject.create();
+  assert.ok(subject);
+});
